refactor: separate question edit modal into components
